started adding qr code option
added default alphanumeric code on new class generation, so that one of the radio buttons will be checked. The checked one will be added to the db under new field "verification method" under that class. Made it so the textfield that displays the alphanumeric code will only show if that class uses alphanumeric method of check in.